Split up settings/utils to break circular dependency chain. New utils modules are in a settings/utils dir with an __init__.py, so old imports still resolve
