+++ phac_aspc/django/settings/utils/__init__.py
@@ -1,0 +1,5 @@
+# flake8: noqa
+"""Settings utils module"""
+from .configuration_utils import *
+from .env_utils import *
+from .test_utils import *
